Added the basic interface and main files.
Added the Component and System interface files. These will be used to
create systems and components with a standard interface to expose to
scripting. However systems are allowed to interact with components
however they see fit.

The main file is polluted with some test code that will be removed
within the next few commits. win32 is just some basic win32 window
creation code and will be modularized in the OS independence branch.

Entity is included but will have its usage and named changed to reflect
the fact that it is a property class.