Lots of platform agnostic changes.
IOpSys now creates the OpenGL context during CreateGraphicsWindow().
Also there is now a check from the return of Update to verify if an
update interval passed. If the check was true, then Present is called
from IOpSys to swap the front and back buffer.

Minor fix for an include in GLSLShader. Linux is case-sensitive in its
directories.