Implemented the camera controller
The camera controller is now implemented. The current method though is
lacking in smooth movement, and is borked when pressing a second key.
The behavior is correct and not a bug. What is needed is a physics
system to apply a movement on update and the event system will then need
to add or remove a constant force instead of adding a constant movement
each event.